requirements.txt
python-telegram-bot
pyyaml
asyncio
pillow